fix a bug where a new user cannot open the My Created Events screen
created a new user, adjusted some lines of code because they are unable to create an event, add a new attribute in the appwrite console to successfully create an event. New user successfully created an event. Logged in with Hello World user and they are unable to see the new user's created event. (Successful)